edit pages routing path, add columns props in table components
